XMP-77: rename Context;
