+++ xampl/Models/Xampl/User.cs
@@ -3,7 +3,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace xampl.Models.Documents;
+namespace xampl.Models.Xampl;
 
 /// <summary>
 /// Store main user data
